renamed NameValuePair to NameValuePairImpl due to google's naming convention renamed INameValuePair to NameValuePair due to google's naming convention
