--- src/main/java/com/github/gwtbootstrap/client/ui/incubator/INameValuePair.java
@@ -1,11 +1,0 @@
-package com.github.gwtbootstrap.client.ui.incubator;
-
-/**
- * User: Halil Karakose
- * Date: 10/18/13
- * Time: 6:40 PM
- */
-public interface INameValuePair {
-    String name();
-    String value();
-}
