+++ src/main/java/com/github/gwtbootstrap/client/ui/incubator/NameValuePair.java
@@ -3,24 +3,9 @@ package com.github.gwtbootstrap.client.ui.incubator;
 /**
  * User: Halil Karakose
  * Date: 10/18/13
- * Time: 6:41 PM
+ * Time: 6:40 PM
  */
-public class NameValuePair implements INameValuePair {
-    private String name;
-    private String value;
-
-    public NameValuePair(String name, String value) {
-        this.name = name;
-        this.value = value;
-    }
-
-    @Override
-    public String name() {
-        return name;
-    }
-
-    @Override
-    public String value() {
-        return value;
-    }
+public interface NameValuePair {
+    String name();
+    String value();
 }
